+++ app/controllers/adverse_events_controller.rb
@@ -1,3 +1,4 @@
+# Tracks updates to adverse events
 class AdverseEventsController < ApplicationController
   before_action :authenticate_user!