+++ app/controllers/users_controller.rb
@@ -1,8 +1,10 @@
+# Allows users to set settings, search for other users
+# Allows admins to review existing accounts
 class UsersController < ApplicationController
   before_action :authenticate_user!
-  before_action :check_system_admin, only: [ :new, :create, :edit, :update, :destroy ]
-  before_action :set_user, only: [ :show, :edit, :update, :destroy ]
-  before_action :redirect_without_user, only: [ :show, :edit, :update, :destroy ]
+  before_action :check_system_admin, only: [:new, :create, :edit, :update, :destroy]
+  before_action :set_user, only: [:show, :edit, :update, :destroy]
+  before_action :redirect_without_user, only: [:show, :edit, :update, :destroy]
 
   def settings
   end