+++ db/migrate/20151001134230_create_adverse_event_users.rb
@@ -1,0 +1,11 @@
+class CreateAdverseEventUsers < ActiveRecord::Migration
+  def change
+    create_table :adverse_event_users do |t|
+      t.integer :adverse_event_id
+      t.integer :user_id
+      t.datetime :last_viewed_at
+    end
+
+    add_index :adverse_event_users, [:adverse_event_id, :user_id], unique: true
+  end
+end