+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Project editors can add comments and open and close AEs
   - Files can be attached to an adverse event
   - Sheets can be filled out and added to an adverse event
+  - Users receive a notification if they have viewable adverse events that have new updates, comments, files, or forms
 - **Gem Changes**
   - Updated to pg 0.18.3
   - Removed minitest-reporters